+++ src/luarocks/fs.lua
@@ -42,35 +42,4 @@ fs.verbose = function()    -- patch io.popen and os.execute to display commands 
 end
 if cfg.verbose then fs.verbose() end
 
-local function load_fns(fs_table)
-   for name, fn in pairs(fs_table) do
-      if not fs[name] then
-         fs[name] = fn
-      end
-   end
-end
-
--- Load platform-specific functions
-local loaded_platform = nil
-for _, platform in ipairs(cfg.platforms) do
-   local ok, fs_plat = pcall(require, "luarocks.fs."..platform)
-   if ok and fs_plat then
-      loaded_platform = platform
-      load_fns(fs_plat)
-      break
-   end
-end
-
--- Load platform-independent pure-Lua functionality
-local fs_lua = require("luarocks.fs.lua")
-load_fns(fs_lua)
-
--- Load platform-specific fallbacks for missing Lua modules
-local ok, fs_plat_tools = pcall(require, "luarocks.fs."..loaded_platform..".tools")
-if ok and fs_plat_tools then
-   load_fns(fs_plat_tools)
-   load_fns(require("luarocks.fs.tools"))
-end
-
-
 return fs
