+++ src/luarocks/fs/unix.lua
@@ -132,4 +132,14 @@ function unix.tmpname()
    return os.tmpname()
 end
 
+local function load_fns(fs_table)
+   for name, fn in pairs(fs_table) do
+      if not fs[name] then
+         fs[name] = fn
+      end
+   end
+end
+
+load_fns(unix)
+
 return unix
