+++ src/luarocks/fs/unix/tools.lua
@@ -234,4 +234,14 @@ function tools.make_temp_dir(name)
    return nil, "Failed to create temporary directory "..tostring(dirname)
 end
 
+local function load_fns(fs_table)
+   for name, fn in pairs(fs_table) do
+      if not fs[name] then
+         fs[name] = fn
+      end
+   end
+end
+load_fns(tools)
+
+
 return tools
