+++ src/luarocks/loader.lua
@@ -22,27 +22,9 @@ local manif_core = require("luarocks.manif_core")
 local deps = require("luarocks.deps")
 local util = require("luarocks.util")
 
--- Workaround for wrappers produced by older versions of LuaRocks
+-- As we bundle luarocks inside Tarantool, we don't need to workaround
+-- older LuaRocks
 local temporary_global = false
-if rawget(_G, 'luarocks') then
-   -- The site_config.lua file generated by old versions uses module(),
-   -- so it produces a global `luarocks` table. Since we have the table,
-   -- add the `loader` field to make the old wrappers happy.
-   luarocks.loader = loader
-else
-   -- When a new version is installed on top of an old version,
-   -- site_config.lua may be replaced, and then it no longer creates
-   -- a global.
-   -- Detect when being called via -lluarocks.loader; this is
-   -- most likely a wrapper.
-   local info = debug.getinfo(2, "nS")
-   if info.what == "C" and not info.name then
-      luarocks = { loader = loader }
-      temporary_global = true
-      -- For the other half of this hack,
-      -- see the next use of `temporary_global` below.
-   end
-end
 
 loader.context = {}
